add several mechanics
Added non-stackable aura system
Added coin system
Added delay before first wave
Upgraded tower selector, added sell button
Implemented Slowing system for ice archer
Added end spot for life check
Added comments

+++ Assets/Scripts/Towers/TargetingSystem.cs
@@ -5,7 +5,7 @@ using System.Linq;
 
 public class TargetingSystem : MonoBehaviour
 {
-    private int range;
+    private float range;
     private Tower tower;
     public GameObject currentTarget;
     private void Awake()
